+++ public/favorite.php
@@ -3,7 +3,7 @@
 namespace openSB;
 
 $rawOutputRequired = true;
-require dirname(__DIR__) . '/private/class/common.php';
+require_once dirname(__DIR__) . '/private/class/common.php';
 
 if ($userbandata) {
     error(403, __("Banned user, can't continue."));