+++ public/upload.php
@@ -3,7 +3,7 @@
 namespace openSB;
 
 //this uploads and converts the video, should switch to a better solution!
-require dirname(__DIR__) . '/private/class/common.php';
+require_once dirname(__DIR__) . '/private/class/common.php';
 
 $supportedVideoFormats = ["mp4", "mkv", "wmv", "flv", "avi", "mov", "3gp"];
 $supportedImageFormats = ["png", "jpg", "jpeg"];