+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# Qobo Beta 3
+# openSB Beta 3
 ## Beta 3.0
 ### Beta 3.0.0-003
 * Refinements to Finalium user interface.
 * Replaced Bootstrap icon set with... uh, I forgot.
+* Dropped PHP 7.3 support due to Mobile Detection library.
 ### Beta 3.0.0-002 (November 17th 2022)
 * Finalized image uploading
 * Improvements to Finalium's SCSS